add (ft_set_flags.c/2), finish 0 edge case

+++ ft_itoa_base.c
@@ -6,32 +6,33 @@
 /*   By: kclaes <kclaes@student.codam.nl>             +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/05/08 22:40:21 by kclaes        #+#    #+#                 */
-/*   Updated: 2025/05/10 15:45:03 by kclaes        ########   odam.nl         */
+/*   Updated: 2025/05/11 14:33:00 by kclaes        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft/libft.h"
 #include "ft_printf.h"
 #include <stdlib.h>
+#include <stdint.h>
 
 char	*ft_itoa_base_signed(int val, char *base, bool *sign);
-char	*ft_itoa_base_unsigned(unsigned int val, char *base);
+char	*ft_itoa_base_unsigned(uintptr_t val, char *base);
 
 char	*ft_itoa_base_signed(int val, char *base, bool *sign)
 {
-	unsigned int	uval;
+	uintptr_t	uval;
 
 	if (val < 0)
 	{
-		uval = ((unsigned int) -(val + 1)) + 1;
+		uval = ((uintptr_t) -(val + 1)) + 1;
 		*sign = true;
 	}
 	else
-		uval = (unsigned int) val;
+		uval = (uintptr_t) val;
 	return (ft_itoa_base_unsigned(uval, base));
 }
 
-char	*ft_itoa_base_unsigned(unsigned int val, char *base)
+char	*ft_itoa_base_unsigned(uintptr_t val, char *base)
 {
 	char	buffer[35];
 	char	*str;
@@ -55,56 +56,56 @@ char	*ft_itoa_base_unsigned(unsigned int val, char *base)
 // {
 // 	bool	sign = false;
 // 	char	*s;
-// 	// BINARY 01
-// 	s = ft_itoa_base_signed(0, "01", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
-// 	s = ft_itoa_base_signed(1, "01", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
-// 	s = ft_itoa_base_signed(2, "01", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
-// 	s = ft_itoa_base_signed(10, "01", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
-// 	s = ft_itoa_base_signed(42, "01", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
-// 	s = ft_itoa_base_signed(255, "01", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
-// 	s = ft_itoa_base_signed(-1, "01", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
-// 	s = ft_itoa_base_signed(-42, "01", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
+// 	// // BINARY 01
+// 	// s = ft_itoa_base_signed(0, "01", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
+// 	// s = ft_itoa_base_signed(1, "01", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
+// 	// s = ft_itoa_base_signed(2, "01", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
+// 	// s = ft_itoa_base_signed(10, "01", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
+// 	// s = ft_itoa_base_signed(42, "01", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
+// 	// s = ft_itoa_base_signed(255, "01", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
+// 	// s = ft_itoa_base_signed(-1, "01", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
+// 	// s = ft_itoa_base_signed(-42, "01", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
 // 	//HEX
-// 	s = ft_itoa_base_signed(0, "0123456789ABCDEF", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
-// 	s = ft_itoa_base_signed(10, "0123456789ABCDEF", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
-// 	s = ft_itoa_base_signed(15, "0123456789ABCDEF", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
-// 	s = ft_itoa_base_signed(16, "0123456789ABCDEF", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
-// 	s = ft_itoa_base_signed(31, "0123456789ABCDEF", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
-// 	s = ft_itoa_base_signed(255, "0123456789ABCDEF", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
-// 	s = ft_itoa_base_signed(-4095, "0123456789ABCDEF", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
-// 	s = ft_itoa_base_signed(-255, "0123456789ABCDEF", &sign);
-// 	printf("%s\nsign: %i\n", s, sign);
-// 	free(s);
+// 	// s = ft_itoa_base_signed(0, "0123456789ABCDEF", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
+// 	// s = ft_itoa_base_signed(10, "0123456789ABCDEF", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
+// 	// s = ft_itoa_base_signed(15, "0123456789ABCDEF", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
+// 	// s = ft_itoa_base_signed(16, "0123456789ABCDEF", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
+// 	// s = ft_itoa_base_signed(31, "0123456789ABCDEF", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
+// 	// s = ft_itoa_base_signed(255, "0123456789ABCDEF", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
+// 	// s = ft_itoa_base_signed(-4095, "0123456789ABCDEF", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
+// 	// s = ft_itoa_base_signed(-255, "0123456789ABCDEF", &sign);
+// 	// printf("%s\nsign: %i\n", s, sign);
+// 	// free(s);
 // 	//base 10
 // 	s = ft_itoa_base_signed(0, "0123456789", &sign);
 // 	printf("%s\nsign: %i\n", s, sign);
